pyproject.toml
[tool.poetry]
name = "airflow-custom-operators"
version = "0.1.0"
description = ""
authors = ["davidmorosini <david.morosinineto@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
apache-airflow = "2.1.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
black = "^22.1.0"
autoflake = "^1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
